requirements.txt
beautifulsoup4==4.12.2
chromadb==0.4.18
Flask==3.0.0
flask_sqlalchemy==3.1.1
Flask_User==1.0.2.2
loguru==0.7.2
numpy==1.26.2
pandas==2.1.4
pydantic==2.5.2
python-dotenv==1.0.0
Requests==2.31.0
SQLAlchemy==2.0.23
tqdm==4.66.1
